# Conflicts:
#	homeassistant/components/rest/manifest.json
#	requirements_test_all.txt

+++ homeassistant/components/nad/media_player.py
@@ -1,7 +1,5 @@
 """Support for interfacing with NAD receivers through RS-232."""
 
-from __future__ import annotations
-
 from nad_receiver import NADReceiver, NADReceiverTCP, NADReceiverTelnet
 import voluptuous as vol
 
@@ -214,8 +212,8 @@ class NADtcp(MediaPlayerEntity):
 
     def set_volume_level(self, volume: float) -> None:
         """Set volume level, range 0..1."""
-        nad_volume_to_set = int(
-            round(volume * (self._max_vol - self._min_vol) + self._min_vol)
+        nad_volume_to_set = round(
+            volume * (self._max_vol - self._min_vol) + self._min_vol
         )
         self._nad_receiver.set_volume(nad_volume_to_set)
